test(pnpr): satisfy pacquet ci lint rules
